Loader to parser rename
